+++ assets/js/src/constants/events.js
@@ -1,3 +1,4 @@
 // custom events
 export const AJAX_CART_UPDATE = 'bigcommerce/ajax_cart_update';
 export const HANDLE_CART_STATE = 'bigcommerce/handle_cart_state';
+export const HANDLE_COUPON_CODE = 'bigcommerce/handle_coupon_code';