+++ src/BigCommerce/Plugin.php
@@ -4,7 +4,7 @@
 namespace BigCommerce;
 
 class Plugin {
-	const VERSION = '4.10.0';
+	const VERSION = '4.11.0';
 
 	protected static $_instance;