+++ templates/public/components/cart/cart-footer.php
@@ -4,8 +4,10 @@
  *
  * @package BigCommerce
  *
- * @var string $summary Cart summary including total and taxes.
- * @var string $actions Cart actions.
+ * @var string $shipping    Shipping calculation and sumary.
+ * @var string $coupon_code Coupons calculation and sumary.
+ * @var string $summary     Cart summary including total and taxes.
+ * @var string $actions     Cart actions.
  * @version 1.0.0
  */